Bug 20774: Fix bugs on publisher and subscriber machines
This commit fixes a few bugs that were preventing machines from running
both as subscribers and publishers. One fix was for not requiring a
UserID to be present in the system metadata, as that is an optional
element.  The second fix was to keep track of payload length for journal
files, as that is necessary to re-send the file.