Ticket #5 - Build and run JALoP on RHEL7
Some slight changes were made to the build tooling to allow JALoP to build
with GCC 4.8 and Scons 3.

New warnings supported by GCC required the removal of -Werror for the build
to succeed. _FORTIFY_SOURCE only has an effect when combined with
optimizations, so it is now added only defined in the release build.

Unit tests were failing because test_dept can't replace functions in some
tests when built with optimizations anymore. So, a --no-release option was
added to allow unit testing to be done only on the debug build.

A forward-compatibility shim used in the test_dept build script is no
included in the EL7 python2-scons package and is not needed with Python 2.7.
It is no longer imported.