Ticket #152 - JALoP initialization over HTTP
Initial support for JALoP initialize and initialize-ack messages over HTTP.

jaln_publish and supporting functions were changed to initialize a JALoP connection over HTTP.
New headers were added to configure the digest challenge and to introduce a publisher ID.
The encoding headers were changed to XML compression headers to match the spec.
The dummy_push utility was expanded and used to test with the Java subscriber.