Ticket #301 - jald accepts multiple peers again
jald now takes in the host, port, and initializaion parameters from a list
of peers. It will attempt to establish a connection to each. After a
configurable retry interval, jald will try again to connect if the
initial connection failed or if the connection has been terminated.

Dead code was also cleaned up, and some additional debug logging was added.