Ticket #454 - Fix parsing of sync messages
Sync messages were being improperly parsed, causing the session to
improperly enter the errored state. This led to subsequent records failing
with "early EOF" errors, since the payload was not being attached. A couple
memory leaks and a situation that could lead to a segfault were cleaned up
as well.

Refs #568