Ticket #606 - Handle record failures in response to record
The JNL now treates record-failure messages in response to a sent record
correctly. They can come in place of a digest-challenge or sync, depending
on whether digesting is enabled.

Additionally, the last_message field in jaln_session has been removed in
favor of the one in jaln_response_header_info, and a unit test for
jaln_publisher_send_journal_missing was fixed.