Ticket #536 - Implement close-session
Publisher now sends close-session message on error or jaln_disconnect call.
The on_channel_close and on_connection_close callbacks are executed in most
cases. The code that detects the error or severed connection only get
executed when attempting to send a record, so if there are no records being
sent, the network store will not be notified that it has no connection to the
peer. A heartbeat mechanism will likely need to be added to address this in the
future.