Address hang in send thread after first publish
In newer versions of curl, the function specified by
CURLOPT_READFUNCTION is not called again once the number of bytes
specified by CURLOPT_POSTFIELDSIZE_LARGE is reached. This meant
jaln_pub_feeder_fill_buffer would never get called such that 0 bytes
were handled, and the jaln_pub_feeder_on_finished would never
trigger, and the send thread would wait forever on a signal.

Solved by triggering cleanup on error or the succesful sending of
the payload_break portion of the data, rather than a value of 0
for the byte offset.

Refs #668