Fixes unit tests dealing with XML signatures
The XML dsig spec allows a lot of flexibility in the format of X.509
certificate data. Due to changes in behavior of dependencies, the
elements added in different orders on EL7 and EL8. This commit allows
the unit tests to pass on both platforms. The checks in the unit tests
still make assumptions about the structure, so the unit tests may break
again in the future.